Fix - index the memory buffer using a uint16_t
This is because an opcode is 2 bytes, and it is the type used when fetching instructions from memory

+++ src/Memory.hpp
@@ -43,12 +43,12 @@ namespace chip8
         /// \brief Get the value in the memory location referenced by the given index
         /// \param[in] index The index into the memory block
         /// \returns The value in the memory location at index
-        constexpr uint8_t& operator[](uint8_t index) noexcept;
+        constexpr uint8_t& operator[](uint16_t index) noexcept;
         
         /// \brief Get the value in the memory location referenced by the given index
         /// \param[in] index The index into the memory block
         /// \returns The value in the memory location at index
-        constexpr uint8_t operator[](uint8_t index) const noexcept;
+        constexpr uint8_t operator[](uint16_t index) const noexcept;
 
         /// \brief Load the font set into the memory buffer
         constexpr void loadFontSet() noexcept;
@@ -63,13 +63,13 @@ namespace chip8
         loadFontSet();
     }
 
-    constexpr uint8_t& Memory::operator[](uint8_t index) noexcept
+    constexpr uint8_t& Memory::operator[](uint16_t index) noexcept
     {
         Expects(index >= 0 and index < m_memory.size());
         return m_memory[index];
     }
 
-    constexpr uint8_t Memory::operator[](uint8_t index) const noexcept
+    constexpr uint8_t Memory::operator[](uint16_t index) const noexcept
     {
         Expects(index >= 0 and index < m_memory.size());
         return m_memory[index];
